digital-research-assistant/requirements.txt
aiobotocore==2.19.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioitertools==0.12.0
aiosignal==1.3.2
alembic==1.8.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
asttokens==3.0.0
attrs==23.2.0
audioread==3.0.1
aws-profile-manager==0.7.3
babel==2.16.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
bibtexparser==2.0.0b8
boto3==1.36.3
botocore==1.36.3
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
chromedriver-autoinstaller==0.6.4
click==8.1.3
clldutils==3.24.0
cloudpickle==2.2.1
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
configparser==7.1.0
cryptography==44.0.0
csvw==3.5.1
dataclasses-json==0.6.7
datasets==3.2.0
decorator==5.1.1
dill==0.3.8
distlib==0.3.9
distro==1.9.0
dlinfo==2.0.0
dnspython==2.7.0
docker==7.1.0
executing==2.2.0
fake-useragent==1.5.1
fastapi==0.110.0
filelock==3.17.0
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.1.0
hpack==4.1.0
html2text==2024.2.26
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.27.1
hyperframe==6.1.0
hypothesis==6.124.5
identify==2.6.6
idna==3.10
importlib-metadata==6.11.0
iniconfig==2.0.0
ipinfo==5.1.1
ipython==8.31.0
ipywidgets==8.1.5
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab_widgets==3.0.13
kubernetes==30.1.0
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langsmith==0.1.147
language-tags==1.2.0
lazy_loader==0.4
Levenshtein==0.25.1
librosa==0.10.2.post1
llvmlite==0.44.0
loguru==0.7.3
lxml==5.3.0
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mock==4.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.4.2
nodeenv==1.9.1
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
omegaconf==2.2.3
openai==1.60.1
opik==0.2.2
orjson==3.10.15
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pastel==0.2.1
pathos==0.3.2
pexpect==4.9.0
phonemizer==3.3.0
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
poethepoet==0.29.0
pooch==1.8.2
portalocker==2.10.1
pox==0.3.5
ppft==1.7.6.9
pre-commit==3.8.0
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyctcdecode==0.5.0
pydantic==2.8.2
pydantic-settings==2.7.1
pydantic_core==2.20.1
Pygments==2.19.1
pygtrie==2.5.0
PyJWT==2.7.0
pylatexenc==2.10
pymongo==4.10.1
PyMySQL==1.1.1
pyparsing==3.2.1
PyPDF2==3.0.1
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.13.2
questionary==2.1.0
RapidFuzz==3.11.0
rdflib==7.1.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruff==0.4.10
s3fs==2024.9.0
s3transfer==0.11.2
safetensors==0.5.2
sagemaker==2.237.3
sagemaker-core==1.0.19
sagemaker-huggingface-inference-toolkit==2.4.1
sagemaker_inference==1.10.1
schema==0.7.7
scikit-learn==1.6.1
scipy==1.15.1
secure==0.3.0
segments==2.2.1
selenium==4.28.1
sentence-transformers==3.4.0
six==1.17.0
smdebug-rulesconfig==1.0.1
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
SQLAlchemy==2.0.37
SQLAlchemy-Utils==0.41.2
sqlmodel==0.0.18
stack-data==0.6.3
starlette==0.36.3
sympy==1.13.1
tabulate==0.9.0
tblib==3.0.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tldextract==5.1.3
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.1
trio==0.28.0
trio-websocket==0.11.1
triton==3.1.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uuid7==0.1.0
uvicorn==0.30.6
uvloop==0.21.0
virtualenv==20.29.1
watchfiles==1.0.4
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.2
widgetsnbextension==4.0.13
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
zenml==0.72.0
zipp==3.21.0
